refactor: move executable classes to tasks package
