+++ src/main/kotlin/br/ufpe/liber/tasks/Sitemaps.kt
@@ -1,5 +1,6 @@
-package br.ufpe.liber
+package br.ufpe.liber.tasks
 
+import br.ufpe.liber.Templates
 import br.ufpe.liber.model.BookRepository
 import io.micronaut.context.ApplicationContext
 import io.micronaut.context.env.Environment
